Adapt ConstraintViolationError/ExclusivityViolationError parsing for new EdgeDB logic
Constraints were standardized with: https://github.com/edgedb/edgedb/pull/6796
So now I parse out all the dynamic parts of the message for all constraint errors.

I kept the subclass for now, but it could be replaced with the check
```diff
- e instanceof ExclusivityViolationError
+ e instanceof ConstraintViolationError && e.constraint === 'std::exclusive'
```
