+++ src/core/edgedb/errors/index.ts
@@ -1,0 +1,2 @@
+export * from './constraint-violation.error';
+export * from './exclusivity-violation.error';
